Move the bridge-storage
