+++ packages/apps/src/configs/storages/index.ts
@@ -1,0 +1,2 @@
+export * from './bridge-storage';
+export * from './EvmChainStorage';
